requirements.txt
altair
pandas
streamlit
matplotlib.pyplot
mag4
httpimport